@@ -1,12 +1,11 @@
 apply plugin: 'com.android.application'
 
 android {
-    compileSdkVersion 23
-    buildToolsVersion "23.0.1"
+    compileSdkVersion 28
 
     defaultConfig {
-        minSdkVersion 8
-        targetSdkVersion 23
+        minSdkVersion 14
+        targetSdkVersion 28
         versionName project.VERSION_NAME
         versionCode project.VERSION_CODE
     }
@@ -29,5 +28,6 @@ android {
 }
 
 dependencies {
-    compile project(":library")
+    implementation project(":library")
+    implementation 'androidx.appcompat:appcompat:1.0.2'
 }
